Crate hasp_metadata

Metadata generated by hasp.

Structs

Blake3Hash

A blake3 hash.

CargoDirectory

Specific information associated with Cargo.

DirectoryHash

Represents a directory hash.

DirectoryVersionReq

Version requirement.

InstallFailed

An installation process failed.

InstallInfo

Information about an installation. Returned as part of InstalledPackage.

InstallStarted

A message generated by hasp when an installation is started.

InstallSuccess

An installation process succeeded.

InstalledFile

Represents a binary that is currently installed. Returned as part of InstallInfo.

InstalledPackage

Represents a package that is currently installed.

PackageDirectory

Information about a directory installation for a single package.

ParseDirectoryVersionError

An error encountered while parsing a directory version.

ParseHashError

An error encountered while parsing a hash from a string or SQL blob.

Enums

DirectoryVersion

Represents a directory version.

FailureReason

The reason a package installation failed.

FileHash

A hash for an installed file. Returned as part of InstalledFile.